Added Sprite Animations for Common and Elite Enemies and Updated Sizes of enemies
- There is now 1 enemy animation variation each for common and elite enemies.
- Updated enemy sizes, hitboxes, dash speeds and walk speeds to compensate for the reduced size of the sprites.

+++ Group Project/Assets/Enemy/Enemy Base Components/Enemy.cs
@@ -7,7 +7,7 @@ public abstract class Enemy : MonoBehaviour
     // Generic enemy stats
     protected int health;
     protected int contactDamage;
-    protected int moveSpeed;
+    protected float moveSpeed;
 
     // Generic enemy drops
     protected float expOnDeath;
